EN: Added a default implementation to GetProviderId

+++ iModelCore/ecobjects/src/presentation/auiprovider.cpp
@@ -14,7 +14,7 @@ USING_NAMESPACE_EC
 /*---------------------------------------------------------------------------------**//**
 * @bsimethod                                    Abeesh.Basheer                  04/2012
 +---------------+---------------+---------------+---------------+---------------+------*/
-UInt16          IECPresentationProvider::GetProviderId () const
+size_t          IECPresentationProvider::GetProviderId () const
     {
     return _GetProviderId();
     }
